2023-06-26 :: complete responsive
