+++ package.json
@@ -20,7 +20,7 @@
     "crypto": "^1.0.1",
     "firebase": "^9.10.0",
     "mcm-js": "^0.0.41",
-    "mcm-js-commons": "^0.0.62",
+    "mcm-js-commons": "^0.0.64",
     "mcm-js-dev": "^0.0.27",
     "next": "13.1.6",
     "react": "18.2.0",
