+++ src/main/commonsComponents/units/template/form/comments/list/contents/select/functional/contents.select.functional.styles.ts
@@ -1,10 +1,20 @@
 import styled from "@emotion/styled";
+import { breakPoints } from "mcm-js-commons/dist/responsive";
 
 interface StyleTypes {
   isSuccess?: boolean;
   isDelete?: boolean;
   isReviewCategory?: boolean;
 }
+
+export const Form = styled.form`
+  @media ${breakPoints.mobileLarge} {
+    display: flex;
+    width: 100%;
+    height: 100%;
+    align-items: center;
+  }
+`;
 
 export const OptionalWrapper = styled.div`
   display: flex;
@@ -19,8 +29,6 @@ export const OptionalWrapper = styled.div`
   }
 
   .optional-input {
-    /* margin-top: 30px; */
-
     .mcm-textArea-unit {
       height: 220px;
       resize: none;
@@ -45,6 +53,23 @@ export const OptionalWrapper = styled.div`
   .mcm-close-button-unit {
     border-top: unset;
   }
+
+  @media ${breakPoints.mobileLarge} {
+    padding: 0rem;
+    height: auto;
+
+    .mcm-title-unit {
+      /* margin: 10px 0px; */
+      margin-bottom: 20px;
+    }
+
+    .optional-input {
+      .mcm-textArea-unit {
+        height: 180px;
+        font-size: 12px;
+      }
+    }
+  }
 `;
 
 export const ConfirmButtonWrapper = styled.div`
@@ -63,6 +88,14 @@ export const ConfirmButtonWrapper = styled.div`
   .able {
     color: black;
     cursor: pointer;
+  }
+
+  @media ${breakPoints.mobileLarge} {
+    margin-top: 20px;
+
+    .confirm-button {
+      font-size: 14px;
+    }
   }
 `;
 
@@ -84,6 +117,11 @@ export const Message = styled.div`
 export const CommentsInfoWrapper = styled.div`
   margin: 24px 0px;
   margin-bottom: 10px;
+
+  @media ${breakPoints.mobileLarge} {
+    margin: 0px;
+    margin-bottom: 10px;
+  }
 `;
 
 export const CommentsInfoItems = styled.div`
@@ -107,6 +145,18 @@ export const CategoryInfo = styled.div`
     .stars-wrapper {
       margin-top: 0px;
       width: auto;
+    }
+  }
+
+  @media ${breakPoints.mobileSmall} {
+    .label-wrapper {
+      flex-wrap: wrap;
+      gap: 6px 12px;
+    }
+
+    p {
+      font-size: 10px;
+      padding: 6px;
     }
   }
 `;
