+++ src/main/commonsComponents/units/template/form/comments/write/privacy/index.tsx
@@ -1,5 +1,4 @@
-import styled from "@emotion/styled";
-
+import { AgreeUserPrivacyWrapper, PrivacyItems } from "./privacy.styles";
 import { _Title, _Button } from "mcm-js-commons";
 import { WriteInfoTypes } from "../comments.write.types";
 
@@ -40,10 +39,9 @@ export default function PrivacyPage({
       id: "privacy-notice-modal",
       onFixWindow: true,
       closeMent: "닫기",
-      modalStyles: {
-        contents: {
-          padding: "2rem",
-        },
+      mobileModalSize: {
+        width: "95%",
+        height: "450px",
       },
     });
   };
@@ -76,77 +74,3 @@ export default function PrivacyPage({
     </AgreeUserPrivacyWrapper>
   );
 }
-
-interface StyleTypes {
-  checked?: boolean;
-}
-
-export const AgreeUserPrivacyWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0px 8px;
-
-  #privacy-label {
-    font-size: 14px;
-    cursor: pointer;
-    color: gray;
-    width: 194px;
-  }
-
-  .checked {
-    color: #30a2ff !important;
-    font-weight: 700;
-  }
-
-  .privacy-notice {
-    color: gray;
-  }
-`;
-
-export const PrivacyItems = styled.div`
-  display: flex;
-
-  // 원본 체크박스
-  #privacy-checkbox {
-    display: none;
-    transition: all 0.25s ease;
-  }
-
-  // 원본 체크박스를 대체할 체크박스 스타일
-  #privacy-checkbox-label {
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    border: 2px solid #bcbcbc;
-    border-color: ${(props: StyleTypes) => props.checked && "#00c4ff"};
-    border-radius: 100%;
-    cursor: pointer;
-    transition: all 0.25s ease;
-    position: relative;
-
-    ::after,
-    ::before {
-      position: absolute;
-      content: "";
-      width: 10px;
-      height: 2px;
-      background-color: #00c4ff;
-      opacity: ${(props) => Number(props.checked)};
-    }
-
-    ::after {
-      transform: rotate(45deg);
-      width: 5px;
-      top: 9px;
-      left: 2.5px;
-      border-radius: 50%;
-    }
-
-    ::before {
-      transform: rotate(-45deg);
-      left: 4.5px;
-      top: 7.5px;
-      border-radius: 50%;
-    }
-  }
-`;
