+++ src/main/commonsComponents/units/template/form/comments/write/stars/stars.tooltip.text.ts
@@ -1,0 +1,18 @@
+// 버그 및 리뷰 선택할 경우의 점수 메세지
+
+export const starsTooltipTextList: { [key: string]: Array<string> } = {
+  bug: [
+    "정상 : 매우 소소한 이슈예요.",
+    "보통 : 이슈는 있지만 실행에는 문제 없어요.",
+    "주의 : 가끔 실행에 문제가 있어요.",
+    "경고 : 빈번하게 실행에 문제가 있어요.",
+    "긴급 : 실행이 아예 안되고 있어요.",
+  ],
+  review: [
+    "😡 개선이 많이 필요해요.",
+    "🙁 그냥 그래요.",
+    "🤔 애매해요.",
+    "🙂 쓸만해요.",
+    "😁 너무 좋아요.",
+  ],
+};
